change folder for market, stocks, trade
