+++ src/app/personal-assets/page.tsx
@@ -115,6 +115,30 @@ export default function HomePage() {
     }
   };
 
+  // Handle Sell
+  const handleSell = async (assetName: string) => {
+    try {
+      const response = await fetch('/api/sell', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          'user-id': userId || "", // Pass userId in the headers
+        },
+        body: JSON.stringify({ assetName }),
+      });
+
+      if (!response.ok) {
+        throw new Error("Failed to sell asset");
+      }
+
+      const data = await response.json();
+      setBalance(data.balance); // Update the balance in the state
+      setAssets(data.assets); // Update the assets in the state
+    } catch (error) {
+      setError('Error selling asset');
+    }
+  };
+
   if (loading) {
     return (
       <Layout>
@@ -173,7 +197,7 @@ export default function HomePage() {
               </h2>
               <div className="rounded-lg bg-[#0d1829] p-6">
                 <span className="text-4xl font-bold text-green-400">
-                  ${balance.toLocaleString()}
+                  {balance.toFixed(2)}
                 </span>
               </div>
             </div>
@@ -215,6 +239,7 @@ export default function HomePage() {
                     <th className="py-2 px-4 border-b border-gray-700 text-white">Amount</th>
                     <th className="py-2 px-4 border-b border-gray-700 text-white">Price (ETH)</th>
                     <th className="py-2 px-4 border-b border-gray-700 text-white">Total Value</th>
+                    <th className="py-2 px-4 border-b border-gray-700 text-white">Action</th>
                   </tr>
                 </thead>
                 <tbody>
@@ -224,6 +249,11 @@ export default function HomePage() {
                       <td className="py-2 px-4 border-b border-gray-700 text-white">{asset.quantity}</td>
                       <td className="py-2 px-4 border-b border-gray-700 text-white">{asset.price.toFixed(2)}</td>
                       <td className="py-2 px-4 border-b border-gray-700 text-white">{asset.totalValue.toFixed(2)}</td>
+                      <td className="py-2 px-4 border-b border-gray-700 text-white">
+                        <Button variant="outline" className="text-white" onClick={() => handleSell(asset.name)}>
+                          Sell
+                        </Button>
+                      </td>
                     </tr>
                   ))}
                 </tbody>
