+++ src/app/popular-stocks/loading.tsx
